requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.8.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
boto3==1.36.19
botocore==1.36.19
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.7
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==46.0.6
daphne==4.1.2
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.4
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jmespath==1.1.0
msgpack==1.1.2
packaging==26.0
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
psycopg2-binary==2.9.9
pyOpenSSL==26.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.8
requests==2.33.1
s3transfer==0.11.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
whitenoise==6.7.0
zope.interface==8.2
